Simpler queue based search.
We now have a global channel that akts as a queue for
breadth-first search. If computing the nodes of the tree
is fast, then this should cause significant overhead. But if
the main work of a program is computing the nodes then even this
simple approach might work quite well. It should scale good because
no thread needs to be idle until all work is done.